Fix croppie styles, update function signature.
CropModal now expects a reference to the button element, an ID to use in
the DOM for the modal & associated buttons & handlers, a path to the
image, & a handler for doing whatever is needed with the resulting
cropped image.

Fixed config for building as a library, separate from the demo in
root/index.html. Now exports only the CropModal function & doesn't
include anything from root/index.html or src/main.ts in built lib at
all.